Update imports outside of sync/ to point to newly split files
